requirements.txt
selenium==4.23.0
pytest==8.4.2
pytest-xdist==3.6.1
pytest-html==4.1.1
webdriver-manager==4.0.2
requests==2.32.3
Pillow==10.3.0
